+++ test/test_Matrix.cpp
@@ -68,18 +68,18 @@ protected:
         }
     }
     static void check_element(
-        double elem, int b, int j, int Nt, int Nm, int Nd, bool conj, bool full)
+        double elem, size_t b, size_t j, size_t Nt, size_t Nm, size_t Nd, bool conj, bool full)
     {
         double correct_elem;
         if (conj) {
             if (full) {
-                correct_elem = (Nm * Nd * ((j + 1) * (2 * Nt - j))) / 2;
+                correct_elem = (Nm * Nd * ((j + 1) * (2 * Nt - j))) / 2.0;
             } else {
                 correct_elem = (Nt - j) * Nd;
             }
         } else {
             if (full) {
-                correct_elem = (Nm * Nd * ((Nt - j) * (2 * (j + 1) + (Nt - j - 1)))) / 2;
+                correct_elem = (Nm * Nd * ((Nt - j) * (2 * (j + 1) + (Nt - j - 1)))) / 2.0;
             } else {
                 correct_elem = (j + 1) * Nm;
             }
@@ -106,8 +106,8 @@ protected:
             gpuErrchk(
                 cudaMemcpy(h_vec, d_vec, num_blocks * Nt * sizeof(double), cudaMemcpyDeviceToHost));
 
-            for (int i = 0; i < num_blocks; i++) {
-                for (int j = 0; j < Nt; j++) {
+            for (size_t i = 0; i < num_blocks; i++) {
+                for (size_t j = 0; j < Nt; j++) {
                     check_element(h_vec[i * Nt + j], i, j, Nt, Nm, Nd, conj, full);
                 }
             }
